Support native context.Context when go1.7 is used.
